move testconnectionbadge component to its own file
